feat: dev server with serve static support
