+++ package.json
@@ -58,7 +58,7 @@
     "changelogen": "^0.5.4",
     "eslint": "^8.46.0",
     "eslint-config-unjs": "^0.2.1",
-    "h3": "^1.8.0-rc.2",
+    "h3": "1.8.0-rc.2",
     "ip-regex": "^5.0.0",
     "prettier": "^3.0.0",
     "typescript": "^5.1.6",
