+++ playground/public/test.txt
@@ -1,0 +1,1 @@
+Static asset works!
