+++ src/_utils.ts
@@ -1,9 +1,6 @@
 import { promises as fs } from "node:fs";
 import { networkInterfaces } from "node:os";
-import { relative, resolve } from "node:path";
 import { colors } from "consola/utils";
-import { fileURLToPath } from "mlly";
-import { isAbsolute } from "pathe";
 import type { Certificate, HTTPSOptions } from "./types";
 
 export async function resolveCert(
@@ -68,34 +65,4 @@ export function formatURL(url: string) {
       decodeURI(url).replace(/:(\d+)\//g, `:${colors.bold("$1")}/`),
     ),
   );
-}
-
-export async function createImporter(input: string, _cwd?: string) {
-  const cwd = resolve(_cwd ? fileURLToPath(_cwd) : ".");
-
-  const jiti = await import("jiti").then((r) => r.default || r);
-  const _jitiRequire = jiti(cwd, {
-    esmResolve: true,
-    requireCache: false,
-    interopDefault: true,
-  });
-
-  if (!isAbsolute(input) && !input.startsWith(".")) {
-    input = `./${input}`;
-  }
-
-  const entry = _jitiRequire.resolve(input);
-
-  const _import = () => {
-    const r = _jitiRequire(input);
-    return Promise.resolve(r.default || r);
-  };
-
-  return {
-    cwd,
-    relative: (path: string) => relative(cwd, path),
-    formateRelative: (path: string) => `\`./${relative(cwd, path)}\``,
-    entry,
-    import: _import,
-  };
 }
