+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from "./listen";
 export * from "./types";
-export * from "./watch";
+export * from "./server";
